requirements.txt
Django==1.8.9
djangorestframework==3.3.2